+++ internal/provider/resource_iosxr_router_bgp_neighbor_group.go
@@ -174,6 +174,10 @@ func (r *RouterBGPNeighborGroupResource) Schema(ctx context.Context, req resourc
 							MarkdownDescription: helpers.NewAttributeDescription("Prevent next-hop-self from being inherited from the parent").String,
 							Optional:            true,
 						},
+						"route_reflector_client": schema.BoolAttribute{
+							MarkdownDescription: helpers.NewAttributeDescription("Configure a neighbor as Route Reflector client").String,
+							Optional:            true,
+						},
 						"route_reflector_client_inheritance_disable": schema.BoolAttribute{
 							MarkdownDescription: helpers.NewAttributeDescription("Prevent route-reflector-client from being inherited from the parent").String,
 							Optional:            true,