+++ docs/guides/changelog.md
@@ -13,6 +13,7 @@ description: |-
 - Add support for leaf-lists to `iosxr_gnmi` resource
 - Add `port` and `operator` attributes to `iosxr_logging_vrf` resource and data source
 - Add `iosxr_router_hsrp_interface_address_family_ipv6_group_v2` resource and data source
+- Add `route_reflector_client` to `iosxr_router_bgp_neighbor_address_family` and `iosxr_router_bgp_neighbor_group` resource and data source
 
 ## 0.2.5